requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
httpx==0.25.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
